fix: address PR #1 review feedback
Correctness fixes:
- Fetch relation was mapped to the "root" case name, so fetch relations
  were named "root_N" (colliding with real roots) instead of "fetch_N".
- Converter PipelineVisitor saved the enclosing relation scope in a single
  slot, corrupting the scope on 3+ levels of relation nesting; replaced with
  a stack that restores the correct parent scope per level.

FFI robustness (lib.rs / python / go):
- Marked the FFI entry points `unsafe extern "C"` with safety docs and
  dropped the now-redundant clippy allows.
- Replaced `.expect("Invalid layout")` panics (UB across the FFI boundary)
  with null-pointer / early-return handling.
- Switched the length prefix from `usize` to a fixed-width `u64` and updated
  the Python and Go bindings in lockstep.
- Read the buffer address via `cast(ptr, c_void_p).value` in the Python
  binding instead of `addressof(ptr.contents)`.

Consistency:
- SymbolInfo lock accessors now uniformly `.unwrap()` instead of silently
  swallowing poisoned locks in some setters/getters.

Tests:
- Added regression tests for the fetch case name and nested relation scopes
  (both verified to fail without their respective fixes).

+++ go/substrait/textplan.go
@@ -44,8 +44,8 @@ func (tp *TextPlan) LoadFromText(text string) ([]byte, error) {
 		return nil, errors.New("failed to load text plan")
 	}
 
-	// First sizeof(size_t) bytes contain the length
-	lenPtr := (*C.size_t)(ptr)
+	// First sizeof(uint64) bytes contain the length
+	lenPtr := (*C.uint64_t)(ptr)
 	length := int(*lenPtr)
 
 	// Rest is the data
